TURTLESHROOM PRODUCTIONS PROUDLY PRESENTS: Based Sandvich
Alright, so I was looking through your files, and I've noticed that you made a rookie mistake. Namely, that you made a Meal (or food-medicine in this case) that uses the internal Recipe Maker/Cost List for a Thing Def. It is ALWAYS better to use the Recipe system, because Mods that remove an ingredient from something like that cause the entire Recipe to fail.

For example, I have a Mod that removes the Vanilla Plants Expanded Mod's Cabbage and replaces it with another form. When I try to load your recipe, it causes the game to freak out and spam errors. However, if the Sandwich Medicine was made by a Recipe Def instead, this would not happen.

Except, I'm not going to tell you. I'm going to demonstrate it and make your Mod as good as it can possibly be.